services/ml-service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
torch==2.4.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
psycopg2-binary==2.9.9
prometheus-fastapi-instrumentator==7.0.0
